backend complete with authentication

+++ backend/controllers/goalControllers.js
@@ -1,5 +1,5 @@
 const asyncHandler = require('express-async-handler')
-
+const User = require('../models/userModel')
 const Goal = require('../models/goalModel')
 
 
@@ -7,7 +7,7 @@ const Goal = require('../models/goalModel')
 //route GET /api/goals
 //access private
 const getGoals =asyncHandler( async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     
     res.status(200).json(goals)
 })
@@ -24,6 +24,7 @@ const setGoals = asyncHandler( async (req,res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -38,6 +39,20 @@ const updateGoals = asyncHandler( async (req,res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if (!user) {
+        res.status(401) 
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -59,6 +74,20 @@ const deleteGoals = asyncHandler( async (req,res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    
+    //check for user
+    if (!user) {
+        res.status(401) 
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     
     await goal.deleteOne()
